+++ src/chronos.h
@@ -28,6 +28,6 @@ along with this program. if not, see <https://www.gnu.org/licenses/> */
 void set_localtime(struct cdata *cdata);
 int months(int month, int year);
 void chart_timeset(struct cdata *cdata,int *day_offset);
-void zxx_fill(double *z_arr[]);
-void pxx_fill(double cusps[], double *p_arr[], double *z_arr[],
+void zxx_fill(int *z_arr[]);
+void pxx_fill(double cusps[], double *p_arr[], int *z_arr[],
 struct cdata *cdata, struct pxx *pxx);
